refactor: rename version tab to configuration
